Combining audio files to mono sound
PRAAT Script for combining two audio files into a mono sound. An input .tsv file indicates audio file pairs. This script initiates a search for the matched files in the defined directory. The audio files are normalized to an intensity of 70 dB SPL and are converted to a mono sound.